% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dispatch_granularity}
\alias{dispatch_granularity}
\title{Get the history depending on granularity}
\usage{
dispatch_granularity(callback_clr, granularity)
}
\arguments{
\item{callback_clr}{A callback.}

\item{granularity}{Either "epoch" or "iteration".}
}
\value{
The history in a tabular format.
}
\description{
Get the history depending on granularity
}
